Major impovements: separated tests via files fumctional
